# Use quickpartition for RTree bulk loading