--- a/run_benchmarks.sh
+++ b/run_benchmarks.sh
@@ -14,8 +14,8 @@ old="$(mktemp)"
 new="$(mktemp)"
 trap "rm -f $new $old" EXIT
 
-package="./rtree"
-bench="Bulk"
+package="./..."
+bench="."
 
 for (( i = 0; i < 15; i++ )); do
 	echo