## Description

Increases performance of bulk loading by using quickpartition instead of a regular sort.

We're essentially replacing an O(n) algorithm with an O(nlogn) algorithm.

## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- #227 

## Benchmark Results

```
name                                          old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       2.25µs ± 5%    2.13µs ±34%     ~     (p=0.202 n=13+14)
IntersectsLineStringWithLineString/n=100-4      36.2µs ±12%    27.7µs ±44%  -23.39%  (p=0.000 n=13+13)
IntersectsLineStringWithLineString/n=1000-4      503µs ± 4%     271µs ±12%  -46.11%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=10000-4    8.38ms ± 9%    3.83ms ±15%  -54.32%  (p=0.000 n=13+15)
LineStringIsSimpleZigZag/10-4                   2.15µs ± 7%    2.15µs ±10%     ~     (p=0.822 n=14+15)
LineStringIsSimpleZigZag/100-4                  61.4µs ± 2%    61.8µs ± 4%     ~     (p=0.458 n=13+14)
LineStringIsSimpleZigZag/1000-4                 1.01ms ±10%    1.00ms ± 9%     ~     (p=0.170 n=13+15)
LineStringIsSimpleZigZag/10000-4                13.5ms ± 4%    13.6ms ± 7%     ~     (p=0.239 n=13+14)
PolygonSingleRingValidation/n=10-4              2.50µs ±18%    2.38µs ± 6%     ~     (p=0.087 n=14+15)
PolygonSingleRingValidation/n=100-4             56.3µs ± 5%    58.6µs ± 5%   +4.13%  (p=0.001 n=14+14)
PolygonSingleRingValidation/n=1000-4            1.05ms ± 8%    1.04ms ± 4%     ~     (p=0.720 n=14+13)
PolygonSingleRingValidation/n=10000-4           14.3ms ± 5%    14.4ms ± 8%     ~     (p=0.935 n=15+15)
PolygonMultipleRingsValidation/n=4-4            8.31µs ± 7%    8.42µs ± 7%     ~     (p=0.223 n=13+13)
PolygonMultipleRingsValidation/n=36-4           80.4µs ±12%    78.3µs ± 4%     ~     (p=0.419 n=15+14)
PolygonMultipleRingsValidation/n=400-4          1.02ms ± 2%    1.04ms ± 9%     ~     (p=0.614 n=12+15)
PolygonMultipleRingsValidation/n=4096-4         13.0ms ±15%    12.5ms ± 9%     ~     (p=0.116 n=14+13)
PolygonZigZagRingsValidation/n=10-4             19.5µs ±10%    17.5µs ± 6%  -10.34%  (p=0.000 n=15+13)
PolygonZigZagRingsValidation/n=100-4             292µs ± 5%     263µs ± 5%   -9.78%  (p=0.000 n=12+14)
PolygonZigZagRingsValidation/n=1000-4           4.10ms ± 6%    3.40ms ±15%  -17.17%  (p=0.000 n=14+15)
PolygonZigZagRingsValidation/n=10000-4          55.9ms ± 2%    42.3ms ± 4%  -24.31%  (p=0.000 n=12+15)
MultipolygonValidation/n=1-4                     334ns ± 6%     341ns ± 8%     ~     (p=0.268 n=13+13)
MultipolygonValidation/n=4-4                     883ns ± 6%     881ns ±10%     ~     (p=0.693 n=14+13)
MultipolygonValidation/n=16-4                   5.61µs ± 6%    5.64µs ± 7%     ~     (p=1.000 n=14+15)
MultipolygonValidation/n=64-4                   33.9µs ± 4%    33.5µs ± 3%     ~     (p=0.185 n=14+13)
MultipolygonValidation/n=256-4                   199µs ± 4%     198µs ± 6%     ~     (p=0.401 n=14+14)
MultipolygonValidation/n=1024-4                 1.03ms ±19%    0.98ms ± 7%     ~     (p=0.098 n=15+15)
MultiPolygonTwoCircles/n=10-4                   5.65µs ± 6%    4.41µs ± 8%  -21.84%  (p=0.000 n=13+13)
MultiPolygonTwoCircles/n=100-4                  83.3µs ±12%    47.6µs ± 7%  -42.85%  (p=0.000 n=14+14)
MultiPolygonTwoCircles/n=1000-4                 1.20ms ± 7%    0.51ms ± 7%  -57.17%  (p=0.000 n=14+14)
MultiPolygonTwoCircles/n=10000-4                18.8ms ± 9%     7.6ms ±22%  -59.81%  (p=0.000 n=13+15)
MultiPolygonMultipleTouchingPoints/n=1-4        5.49µs ± 4%    5.30µs ± 9%   -3.58%  (p=0.005 n=14+15)
MultiPolygonMultipleTouchingPoints/n=10-4       44.0µs ± 9%    41.2µs ± 7%   -6.47%  (p=0.000 n=13+14)
MultiPolygonMultipleTouchingPoints/n=100-4       533µs ± 9%     469µs ± 7%  -11.92%  (p=0.000 n=14+15)
MultiPolygonMultipleTouchingPoints/n=1000-4     6.83ms ± 7%    5.63ms ± 7%  -17.55%  (p=0.000 n=13+15)
MarshalWKB/polygon/n=10-4                        199ns ±26%     191ns ± 7%     ~     (p=0.777 n=14+14)
MarshalWKB/polygon/n=100-4                       603ns ±19%     565ns ±17%     ~     (p=0.247 n=15+14)
MarshalWKB/polygon/n=1000-4                     3.69µs ±29%    3.51µs ± 8%     ~     (p=0.274 n=14+12)
MarshalWKB/polygon/n=10000-4                    30.7µs ±31%    29.9µs ±59%     ~     (p=0.354 n=15+14)
UnmarshalWKB/polygon/n=10-4                      330ns ±10%     325ns ± 7%     ~     (p=0.547 n=13+15)
UnmarshalWKB/polygon/n=100-4                     690ns ±16%     719ns ±29%     ~     (p=0.198 n=14+13)
UnmarshalWKB/polygon/n=1000-4                   4.31µs ±35%    3.97µs ±18%     ~     (p=0.118 n=14+12)
UnmarshalWKB/polygon/n=10000-4                  32.8µs ±24%    35.9µs ±32%     ~     (p=0.202 n=13+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             48.3µs ± 7%    47.0µs ± 7%     ~     (p=0.098 n=15+15)
Intersection/n=100-4                            91.8µs ± 3%    94.3µs ± 7%     ~     (p=0.116 n=13+14)
Intersection/n=1000-4                            402µs ± 9%     400µs ±12%     ~     (p=0.769 n=14+14)
Intersection/n=10000-4                          3.38ms ±10%    3.58ms ±22%     ~     (p=0.102 n=14+15)
NoOp/n=10-4                                     3.76µs ± 4%    3.73µs ± 6%     ~     (p=0.488 n=13+14)
NoOp/n=100-4                                    12.5µs ± 9%    12.5µs ± 6%     ~     (p=0.905 n=14+13)
NoOp/n=1000-4                                   89.3µs ± 6%    88.0µs ± 7%     ~     (p=0.354 n=15+14)
NoOp/n=10000-4                                   929µs ±23%     905µs ±35%     ~     (p=0.305 n=15+15)
pkg:github.com/peterstace/simplefeatures/rtree goos:linux goarch:amd64
Delete/n=100-4                                  30.8µs ± 2%    31.6µs ± 4%   +2.58%  (p=0.004 n=15+13)
Delete/n=1000-4                                  745µs ± 9%     749µs ± 6%     ~     (p=0.186 n=15+14)
Delete/n=10000-4                                34.1ms ± 3%    33.8ms ± 4%   -0.95%  (p=0.029 n=14+15)
Bulk/n=10-4                                     1.43µs ±29%    0.96µs ±58%  -32.92%  (p=0.000 n=15+14)
Bulk/n=100-4                                    33.3µs ± 9%    14.4µs ±11%  -56.77%  (p=0.000 n=15+15)
Bulk/n=1000-4                                    743µs ± 6%     224µs ± 6%  -69.80%  (p=0.000 n=15+15)
Bulk/n=10000-4                                  13.1ms ±11%     3.2ms ± 8%  -75.35%  (p=0.000 n=14+15)
Bulk/n=100000-4                                  189ms ± 5%      38ms ± 9%  -80.11%  (p=0.000 n=13+15)
Insert/n=10-4                                   1.53µs ±13%    1.47µs ±17%     ~     (p=0.056 n=14+13)
Insert/n=100-4                                  35.1µs ±12%    34.8µs ± 6%     ~     (p=0.683 n=15+15)
Insert/n=1000-4                                  576µs ± 7%     561µs ± 3%   -2.61%  (p=0.001 n=15+14)
Insert/n=10000-4                                7.87ms ± 6%    7.85ms ± 5%     ~     (p=0.751 n=15+13)
Insert/n=100000-4                                101ms ± 7%     100ms ± 6%     ~     (p=0.285 n=15+15)
RangeSearch/n=10-4                              14.9ns ± 4%    14.8ns ± 5%     ~     (p=0.470 n=14+14)
RangeSearch/n=100-4                             57.2ns ± 2%    57.2ns ± 1%     ~     (p=0.517 n=14+14)
RangeSearch/n=1000-4                             209ns ± 2%     211ns ± 4%     ~     (p=0.417 n=13+14)
RangeSearch/n=10000-4                            726ns ± 2%     740ns ± 4%   +1.92%  (p=0.005 n=13+14)
RangeSearch/n=100000-4                          6.98µs ± 4%    7.09µs ± 4%     ~     (p=0.112 n=14+15)

name                                          old alloc/op   new alloc/op   delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       2.54kB ± 0%    2.44kB ± 0%   -3.79%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=100-4      31.4kB ± 0%    30.4kB ± 0%   -3.16%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=1000-4      213kB ± 0%     205kB ± 0%   -3.83%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=10000-4    2.76MB ± 0%    2.63MB ± 0%   -4.75%  (p=0.000 n=15+14)
LineStringIsSimpleZigZag/10-4                   1.44kB ± 0%    1.44kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                  21.6kB ± 0%    21.6kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                  230kB ± 0%     230kB ± 0%   -0.00%  (p=0.006 n=15+12)
LineStringIsSimpleZigZag/10000-4                2.30MB ± 0%    2.30MB ± 0%     ~     (p=0.890 n=15+13)
PolygonSingleRingValidation/n=10-4              1.47kB ± 0%    1.47kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4             16.2kB ± 0%    16.2kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4             217kB ± 0%     217kB ± 0%     ~     (p=0.586 n=15+15)
PolygonSingleRingValidation/n=10000-4           2.23MB ± 0%    2.23MB ± 0%     ~     (p=0.726 n=14+14)
PolygonMultipleRingsValidation/n=4-4            5.31kB ± 0%    5.31kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4           43.6kB ± 0%    43.6kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4           479kB ± 0%     479kB ± 0%     ~     (p=0.085 n=15+14)
PolygonMultipleRingsValidation/n=4096-4         4.92MB ± 0%    4.92MB ± 0%     ~     (p=0.248 n=15+15)
PolygonZigZagRingsValidation/n=10-4             11.4kB ± 0%    11.1kB ± 0%   -2.52%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=100-4             135kB ± 0%     132kB ± 0%   -2.21%  (p=0.000 n=14+13)
PolygonZigZagRingsValidation/n=1000-4           1.04MB ± 0%    1.02MB ± 0%   -2.35%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=10000-4          12.3MB ± 0%    11.9MB ± 0%   -3.19%  (p=0.000 n=14+13)
MultipolygonValidation/n=1-4                      385B ± 0%      385B ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                      676B ± 0%      676B ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                   3.86kB ± 0%    3.86kB ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                   15.4kB ± 0%    15.4kB ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                  63.1kB ± 0%    63.1kB ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                  259kB ± 0%     259kB ± 0%   -0.00%  (p=0.024 n=14+12)
MultiPolygonTwoCircles/n=10-4                   5.17kB ± 0%    4.98kB ± 0%   -3.71%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=100-4                  56.9kB ± 0%    55.0kB ± 0%   -3.48%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=1000-4                  361kB ± 0%     344kB ± 0%   -4.52%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10000-4                4.87MB ± 0%    4.60MB ± 0%   -5.39%  (p=0.000 n=13+15)
MultiPolygonMultipleTouchingPoints/n=1-4        4.00kB ± 0%    3.94kB ± 0%   -1.60%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10-4       23.0kB ± 0%    22.6kB ± 0%   -1.95%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4       176kB ± 0%     172kB ± 0%   -2.29%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1000-4     2.10MB ± 0%    2.04MB ± 0%   -2.84%  (p=0.000 n=15+15)
MarshalWKB/polygon/n=10-4                         232B ± 0%      232B ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                      1.83kB ± 0%    1.83kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                     16.4kB ± 0%    16.4kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                     164kB ± 0%     164kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       282B ± 0%      282B ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                    1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                   16.5kB ± 0%    16.5kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                   164kB ± 0%     164kB ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             1.19kB ± 0%    1.19kB ± 0%     ~     (all equal)
Intersection/n=100-4                            6.33kB ± 0%    6.33kB ± 0%     ~     (p=0.056 n=15+12)
Intersection/n=1000-4                           55.0kB ± 0%    55.0kB ± 0%     ~     (p=0.235 n=14+12)
Intersection/n=10000-4                           557kB ± 0%     558kB ± 0%     ~     (p=0.941 n=15+15)
NoOp/n=10-4                                       864B ± 0%      864B ± 0%     ~     (all equal)
NoOp/n=100-4                                    5.68kB ± 0%    5.68kB ± 0%     ~     (all equal)
NoOp/n=1000-4                                   49.5kB ± 0%    49.5kB ± 0%     ~     (all equal)
NoOp/n=10000-4                                   492kB ± 0%     492kB ± 0%     ~     (p=0.831 n=15+15)
pkg:github.com/peterstace/simplefeatures/rtree goos:linux goarch:amd64
Delete/n=100-4                                    712B ± 0%      712B ± 0%     ~     (all equal)
Delete/n=1000-4                                 25.5kB ± 0%    26.1kB ± 0%   +2.19%  (p=0.000 n=15+15)
Delete/n=10000-4                                 389kB ± 0%     412kB ± 0%   +5.78%  (p=0.000 n=15+15)
Bulk/n=10-4                                     1.54kB ± 0%    1.45kB ± 0%   -6.22%  (p=0.000 n=15+15)
Bulk/n=100-4                                    20.9kB ± 0%    19.9kB ± 0%   -4.75%  (p=0.000 n=15+15)
Bulk/n=1000-4                                    106kB ± 0%      98kB ± 0%   -7.67%  (p=0.000 n=13+15)
Bulk/n=10000-4                                  1.70MB ± 0%    1.57MB ± 0%   -7.69%  (p=0.000 n=12+15)
Bulk/n=100000-4                                 21.5MB ± 0%    20.4MB ± 0%   -4.88%  (p=0.000 n=14+15)
Insert/n=10-4                                   1.44kB ± 0%    1.44kB ± 0%     ~     (all equal)
Insert/n=100-4                                  13.5kB ± 0%    13.5kB ± 0%     ~     (all equal)
Insert/n=1000-4                                  132kB ± 0%     132kB ± 0%     ~     (p=0.070 n=15+13)
Insert/n=10000-4                                1.34MB ± 0%    1.34MB ± 0%     ~     (p=0.195 n=15+15)
Insert/n=100000-4                               13.5MB ± 0%    13.5MB ± 0%     ~     (p=0.107 n=12+15)
RangeSearch/n=10-4                               0.00B          0.00B          ~     (all equal)
RangeSearch/n=100-4                              0.00B          0.00B          ~     (all equal)
RangeSearch/n=1000-4                             0.00B          0.00B          ~     (all equal)
RangeSearch/n=10000-4                            0.00B          0.00B          ~     (all equal)
RangeSearch/n=100000-4                           0.00B          0.00B          ~     (all equal)

name                                          old allocs/op  new allocs/op  delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4         13.0 ± 0%      10.0 ± 0%  -23.08%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=100-4         105 ± 0%        74 ± 0%  -29.52%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=1000-4        601 ± 0%       346 ± 0%  -42.43%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=10000-4     9.56k ± 0%     5.47k ± 0%  -42.83%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/10-4                     5.00 ± 0%      5.00 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                    75.0 ± 0%      75.0 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                    797 ± 0%       797 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000-4                 8.00k ± 0%     8.00k ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10-4                6.00 ± 0%      6.00 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4               57.0 ± 0%      57.0 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4               755 ± 0%       755 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10000-4            7.73k ± 0%     7.73k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4-4              29.0 ± 0%      29.0 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4              239 ± 0%       239 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4           2.63k ± 0%     2.63k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4096-4          26.9k ± 0%     26.9k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10-4               49.0 ± 0%      40.0 ± 0%  -18.37%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=100-4               471 ± 0%       378 ± 0%  -19.75%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=1000-4            3.42k ± 0%     2.66k ± 0%  -22.34%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=10000-4           44.9k ± 0%     32.6k ± 0%  -27.34%  (p=0.000 n=12+12)
MultipolygonValidation/n=1-4                      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                     27.0 ± 0%      27.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                     99.0 ± 0%      99.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                     392 ± 0%       392 ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                  1.58k ± 0%     1.58k ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10-4                     32.0 ± 0%      26.0 ± 0%  -18.75%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=100-4                     216 ± 0%       154 ± 0%  -28.70%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=1000-4                  1.21k ± 0%     0.70k ± 0%  -42.22%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10000-4                 19.1k ± 0%     10.9k ± 0%  -42.82%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-4          49.0 ± 0%      47.0 ± 0%   -4.08%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10-4          308 ± 0%       294 ± 0%   -4.55%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4       2.74k ± 0%     2.61k ± 0%   -4.60%  (p=0.000 n=15+14)
MultiPolygonMultipleTouchingPoints/n=1000-4      28.6k ± 0%     26.7k ± 0%   -6.52%  (p=0.000 n=15+13)
MarshalWKB/polygon/n=10-4                         6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                        6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                     7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                    7.00 ± 0%      7.00 ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                               31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=100-4                              31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=1000-4                             31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=10000-4                            31.0 ± 0%      31.0 ± 0%     ~     (all equal)
NoOp/n=10-4                                       21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=100-4                                      21.0 ± 0%      21.0 ± 0%     ~     (all equal)
…